pyproject.toml
[build-system]
requires = ["setuptools>=75.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "keylight"
version = "0.1.0"
description = "Screen-to-keyboard RGB sync engine for 24-zone keyboards."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "Proprietary" }
authors = [{ name = "KeyLight Team" }]
dependencies = []

[project.optional-dependencies]
hw = [
  "hidapi>=0.15.0",
]
capture = [
  "mss>=10.1.0",
]
dev = [
  "mypy>=1.15.0",
  "pytest>=8.3.0",
  "ruff>=0.9.0",
]

[project.scripts]
keylight = "keylight.cli:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
python_version = "3.12"
strict = true
mypy_path = "src"
packages = ["keylight"]
